Move GetKey into eqxi (to avoid duplication in downstream libraries).
